// aiot-frontend/src/api/generated/model/productUpdateProductRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nunu Example API
 * This is a sample server celler server.
 * OpenAPI spec version: 1.0.0
 */

export interface ProductUpdateProductRequest {
  accessProtocol?: string;
  category?: string;
  connectivityMethod?: string;
  description?: string;
  metadata?: number[];
  name?: string;
  nodeType?: string;
  status?: string;
}